AP_HAL_Linux: renamed AnalogIn_Navio
