+++ libraries/AP_HAL_Linux/AnalogIn_Navio.h
